move to cli package, encapsulation
